chore: reorganize the quarantined 5.9 multiSelect tests into a new root folder
