+++ quarantine/5-9-multiselect-tests/apps-shared/menu-interaction.multiselect.suite.ts
